Refactor into a single story
